fix(gitlab collection): fix pagination, table types, mongo names
1. There was an issue with the fetchPaginated where we were using `?all=true`. I removed that since
it was not giving the correct number of issues in the default branch in gitlab. If you use all=true,
it seems to get about double the commits. I have a theory that this is caused by teams that use
rebase in gitlab. Gitlab seems to save the commits that were rebased with a diffdrent parent so the
ids of the commmits are different. see #95