Run config-ui nginx as non-root
config-ui nginx was running as root, which is not a best practice, as an attacker can potentially become root in a given kubernetes node using some security breach. This commit switches to use an unprivileged nginx container, which runs as nginx user. In addition, it fixes the tag of the image to inherit from, as using latest is also not a good practice, as it can lead to introduce breaking changes if the parent image moves the latest tag to a new image with said breaking changes.