refactor: Use tool/domain scopes as input/output of make-pipeline command
Use the genuine tool scopes as input and genuine domain scopes as output to the make-pipeline command.
For tool scope inputs, python plugins are oblivious of the BlueprintScope concept, the go side taking care
of fetching and passing the real tool scopes.
For the domain scope outputs, it requires to send back the domain scope as a json object together with a type name to allow reconstruction on go side.
Because of limited go reflective abilities, the reconstruction function hardcodes the existing domain scope types.