feat: Support remote scope groups
Python plugins now support a two level scope hierarchy: scopes and scope groups.
When the group_id query parameter is absent, the default implementation calls
`remote_scope_groups` that is redefined in each plugin. If the group_id is given,
the default implementation call `remote_scopes` that return the scopes that are in a given group id.