feat: Better error messages for malformed JSON
The "fire" module parses JSON command arguments as python dict or list.
But when the JSON argument is malformed, fire will just pass a str, silencing any decoding error.
So each time we expect a dict or a list, we reparse the potentially invalid JSON data to get and return the decoding error.