fix: Resolve refs of DynamicModelInfo JSON schema
If a plugin connection, scope or transformation rule has a field with a non-primitive type, pydantic will generate a JSON schema with a reference to that field type.
On go side, the generation of a DynamicTabler from a JSON schema doesn't support JSON refs.
So we need to resolve those refs on python side (thanks to jsonref library) and add support for 'object' JSON type, which is converted to a JSONMap field in the generated struct.