fix: Fix wrong package name in domain ids
RootPkgPath is used by DomainIdGenerator to find the name of the plugin that defines a given type.
While remote plugins do not use the DomainIdGenerator, we still need to implement this function.
Indeed, DomainIdGenerator uses FindPluginNameBySubPkgPath that returns the name of the first plugin
whose RootPkgPath is a prefix of the type package path.
So remote plugin must not return an empty string as RootPkgPath, but a string that is not a prefix
of any go plugin package path.