fix: Fix azure devops migration from 0.17 to 0.18
Cherry-pick of #5957

* fix: Fix AzDO connection migration 0.17 -> 0.18

In 0.17, all connection fields are encrypted.
In 0.18 only `token` is, so we need a migration of AzDo connection to decrypt all fields but `token`.

* fix: Allow setting entities=None to ScopeConfig

Allow setting `entities` field to `None` in `ScopeConfig` and fallback to default value.

* feat: Add RenameColumn migration op

* fix: Add missing AzDO migrations from 0.17 to 0.18

Now that migrations have been made manual, we need to add all the migrations for
all that changed from 0.17 to 0.18 that were automatic before...

Those are all required migrations collected with `git diff v0.17.0 v0.18.0-beta7 -- backend/python/plugins/azuredevops/azuredevops/models.py`:
* Rename GitRepository.transformation_rule_id -> scope_config_id
* Add GitRepository.provider
* Rename Job.startTime -> start_time
* Rename Job.finishTime -> start_time

---------

Co-authored-by: Camille Teruel <camille.teruel@meri.co>