feat: add organization as an optional parameter in Azure DevOps connection form (#7459)
* feat: add organization as an optional parameter in Azure DevOps connection creation.

Users can now create an Azure DevOps connection using a PAT that is limited to a single organization. This means users are not required to use PATs with access to all their organizations.

* feat: add organization as an optional parameter in Azure DevOps connection creation.

Users can now create an Azure DevOps connection using a PAT that is limited to a single organization. This means users are not required to use PATs with access to all their organizations.

* fix: adjust the test data to fix the failed unit test

The test case was incorrect. When an invalid PAT is provided, the Azure DevOps API redirects the user to the login page. Go's HTTP client automatically follows the redirect when it receives a 302 status code. To circumvent this, I made adjustments to the test case.

* fix: TestExistingConnection now incorporates the updated values as well

* chore: replace custom merge logic

---------

Co-authored-by: Klesh Wong <zhenmian.huang@merico.dev>