feat(azuredevops): support disabled Repos (#7657)
* feat(azuredevops): support disabled Azure DevOps Repos

It is now possible to select disabled repositories as a remote scope. However, disabled repositories are not included in tasks related to domain `code` and `code review`, as the Azure DevOps API returns a 4xx status code for the required API calls.

A new column called `is_disabled` has been added to the `_tool_azuredevops_go_repos` table to support disabled repositories. For existing entries, the default value for this column is set to `false`.

* fix(azuredevops): harmonize the DomainTypes of the plugin tasks

* feat(azuredevops): add support for disabled Azure DevOps repos

* fix(azuredevops): fix inconsistent file name

* fix(azuredevops): revert empty entities logic

Empty entities are not treated as 'Selected All' since collecting a scope without any entity is pointless

* fix(azuredevops): fix broken test

* fix(azuredevops): add support for domain type cross

---------

Co-authored-by: Lynwee <linwei.hou@merico.dev>