fix(plugins): add missing scope-config and missing projects in grafana for argocd (#8642)
The scope-config was missing in the initial implementation, which prevent the DORA metrics to be collected.

In Grafana, you could not select the projects, so only ALL was available as an option. I also moved the graphs under the correct sections.

For the comment in the issue https://github.com/apache/incubator-devlake/issues/5207#issuecomment-3537781398